datatype changes

+++ Base64.cpp
@@ -9,7 +9,7 @@ inline void a3_to_a4(unsigned char * a4, unsigned char * a3);
 inline void a4_to_a3(unsigned char * a3, unsigned char * a4);
 inline unsigned char b64_lookup(char c);
 
-int base64_encode(char *output, char *input, int inputLen) {
+double base64_encode(char *output, char *input, int inputLen) {
 	int i = 0, j = 0;
 	int encLen = 0;
 	unsigned char a3[3];
@@ -47,7 +47,7 @@ int base64_encode(char *output, char *input, int inputLen) {
 	return encLen;
 }
 
-int base64_decode(char * output, char * input, int inputLen) {
+double base64_decode(char * output, char * input, int inputLen) {
 	int i = 0, j = 0;
 	int decLen = 0;
 	unsigned char a3[3];
